LPS-124961 Class that implements 'TemplateContextContributor' should be in package *.template
